login and sign-up functionality

+++ php/removetask.php
@@ -1,7 +1,15 @@
 
 <?php 
+    session_start();
     require_once('db.php');
 
+    if (!isset($_SESSION['user_data']))
+    {
+        header('Location: /todolist/login.php');
+        die();
+    }
+
+    $user_id = $_SESSION['user_data']['id'];
     $task_id = $_POST['task_id'];
 
     function remove_task($id)
@@ -18,7 +26,7 @@
         for ($i = 0; $i < $n; $i++)
             $result = $result && remove_task($tasks[$i]['id']);
 
-        $query = "DELETE FROM task WHERE id='$id';";
+        $query = "DELETE FROM task WHERE id='$id' AND user_id='$user_id';";
         $result = $result && run_query($query);
         
         return $result;
